Real Time Post Commenting
Users can now comment on posts via the view post button on the view post page .This stores the users name, avatar image and post body. Sometimes the users image couldn't be pulled so it will store a null instead. Have yet to see this be a problem though if the images start acting funny then that is the reason why